XorCurses
XorCurses is a game based on Xor by Astral Software. XorCurses is written by jwm-art.net, aka, James W. Morris. The original Xor game was released in 1987 for the 8 bit home computers popular at the time. XorCurses attempts to faithfully recreate that game for Linux, with particular attention placed on the behaviour of the objects within the original game.

https://bronevichok.ru/ttygames/#xorcurses ( https://bronevichok.ru/ttygames/#xorcurses )